cyto: convert the simulation chemistry from Float to fixed-point Frac
The sim's conserved quantities now run in the engine's fixed-point Frac (float kept
only at the render / codec / light-field-build boundaries). Two reasons: bulletproof
cross-platform determinism (the AoS↔SoA gate was comparing float bit-identity, brittle
on JS/Native), and exact conservation — the prerequisite for the upcoming closed,
depletable energy system (float drifts; integer + remainder is exactly conserved).

- Normalized to [0,1] (1.0 = a full cell, the old MAX_CHEM): a uniform ÷10 units
  rescale of every energy-dimensioned constant (gate 5→0.5, spawn 2→0.2, threshold
  200→20, light STRENGTH→Frac(1,200), decay constants adjusted). Same dynamics. This
  is also required because Frac×Frac overflows once the product of the two values
  exceeds ~2, so chemistry must live in Frac's [0,1]-ish design range.
- energy/chemicals/transfers/divideCharge/touch/light, gene weights+biases, the
  decay/growth/diffusion/reaction math, and the exposure weight are all Frac now.
  Diffusion floors v/n and the cell keeps the remainder (conservative until the
  depletable grid gives waste somewhere to go). Frac gained sqrt/remainder/fromFloat/
  coerceIn/compareTo for this (committed separately).
- SoA columns store Frac.raw (LongArray); the equivalence gate now compares raw longs.

Verified: :demos:cyto:jvmTest 17/17 (incl. CytoSoaEquivalenceTest bit-identity);
probeCytoPopulation still plateaus (8 cells, flat 2000→8000 ticks).

Still Float, out of this (chemistry) scope — flagged for a follow-up to fully de-float
the sim: spring stress-damage (connection state), the division split-direction geometry
(dot-product ahead/side test), and the light-field bilinear *sample* (Double; becomes
conservation-critical when the field goes depletable).

+++ demos/cyto/src/commonMain/kotlin/org/emerge/demo/cyto/CytoRenderer.kt
@@ -75,7 +75,7 @@ class CytoRenderer {
             for (gx in 0 until FRES) {
                 val wx = -CytoLightField.HALF + (gx + 0.5f) * cell
                 fieldCx[i] = wx; fieldCy[i] = wy
-                val t = (field.sampleAt(wx, wy) / CytoLightField.STRENGTH).coerceIn(0f, 1f)
+                val t = (field.sampleAt(wx, wy).toFloat() / CytoLightField.STRENGTH.toFloat()).coerceIn(0f, 1f)
                 val b = i * 4
                 fieldColor[b] = 0.06f + t * 0.94f
                 fieldColor[b + 1] = 0.05f + t * 0.85f
